External package file loading
I've been thinking a while about this API method and I think could be better to load the package from the "outside". Right now you need to transfer the file to the host before. I've not tried this proposal yet, but even knowing it's really risky I think it could be quite more practical.