Update vst-install-rhel.sh
Not entirely sure what this function is for but it's broken in the current installer. This fixes it. However, on my Centos 7 Install, spamassassin creates it's own /var/lib/spamassassin. So I'm not sure why the script has to make that directory when the yum package manager should take care of that on it's own.