Fix SSL certificates for IDN domains with IDN aliases
Example (real) situation that is being worked around:

IDN domain думалогия.рф has an only alias www.думалогия.рф.
Both domain and its aliases are not punycode-encoded,
that's OK. But they are in punycode in Let's encrypt
certificates. So we have to do so that the check that the
domain alias exists does not fail due to that
'думалогия.рф' != 'xn--80agbsneq0b4h.xn--p1ai'.

The error text was:
"Error: domain XXXX doesn't exist"

Fixes #1809 for IDN domains