Update url grep to https:// to prevent matching on token
I came across the issue where the token contains `url` in the string,
which leads to grep matching on both url and token causing the
`query_le_v2` to fail because the first parsed argument is the url
which is separated by a newline and the token. This eventually leads to
error 15 while requesting a certificate. After some debugging I found
what the issue was. Therefore I'm proposing to make the grep more strict
on `https://`.